# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AssessmentReceipt": {
        "properties": {
          "analysis": {
            "$ref": "#/components/schemas/app__schemas__assessments__AnalysisSummary",
            "description": "Current persisted analysis availability"
          },
          "assessment_id": {
            "format": "uuid",
            "title": "Assessment Id",
            "type": "string"
          },
          "quality": {
            "$ref": "#/components/schemas/QualitySummary"
          },
          "received_at": {
            "description": "UTC ISO-8601 ending in Z",
            "format": "date-time",
            "title": "Received At",
            "type": "string"
          },
          "replayed": {
            "title": "Replayed",
            "type": "boolean"
          },
          "saved": {
            "const": true,
            "default": true,
            "title": "Saved",
            "type": "boolean"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "assessment_id",
          "session_id",
          "received_at",
          "quality",
          "analysis",
          "replayed"
        ],
        "title": "AssessmentReceipt",
        "type": "object"
      },
...
      "AssessmentSubmission": {
        "additionalProperties": false,
        "properties": {
          "attention": {
            "$ref": "#/components/schemas/AttentionRecord"
          },
          "context": {
            "$ref": "#/components/schemas/ContextCheckinIn"
          },
          "memory": {
            "$ref": "#/components/schemas/MemoryRecord"
          },
          "practice": {
            "$ref": "#/components/schemas/PracticeRecord"
          },
          "protocol_version": {
            "maxLength": 64,
            "title": "Protocol Version",
            "type": "string"
          },
          "reaction": {
            "$ref": "#/components/schemas/ReactionRecord"
          },
          "run_duration_ms": {
            "maximum": 1200000.0,
            "minimum": 0.0,
            "title": "Run Duration Ms",
            "type": "number"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "submission_key": {
            "format": "uuid",
            "title": "Submission Key",
            "type": "string"
          },
          "telemetry": {
            "$ref": "#/components/schemas/Telemetry"
          }
        },
        "required": [
          "submission_key",
          "run_id",
          "protocol_version",
          "run_duration_ms",
          "memory",
          "attention",
          "reaction",
          "practice",
          "telemetry",
          "context"
        ],
        "title": "AssessmentSubmission",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "InputMode": {
        "enum": [
          "keyboard",
          "pointer"
        ],
        "title": "InputMode",
        "type": "string"
      },
...
      "PracticeRecord": {
        "additionalProperties": false,
        "properties": {
          "completed": {
            "title": "Completed",
            "type": "boolean"
          },
          "repeats": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Repeats",
            "type": "integer"
          }
        },
        "required": [
          "completed",
          "repeats"
        ],
        "title": "PracticeRecord",
        "type": "object"
      },
...
      "SessionRecoveryResponse": {
        "description": "Owned session metadata. Never task words or raw answers (reload recovery = abandon/new).",
        "properties": {
          "assessment_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessment Id"
          },
          "expires_at": {
            "description": "UTC ISO-8601 ending in Z",
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "input_mode": {
            "$ref": "#/components/schemas/InputMode"
          },
          "protocol_version": {
            "title": "Protocol Version",
            "type": "string"
          },
          "receipt_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Receipt Path"
          },
          "schedule": {
            "$ref": "#/components/schemas/SessionSchedule"
          },
          "scoring_version": {
            "title": "Scoring Version",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "started_at": {
            "description": "UTC ISO-8601 ending in Z",
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "description": "Effective status at server time (expiry is not a GET write)",
            "enum": [
              "STARTED",
              "SUBMITTED",
              "ABANDONED",
              "EXPIRED"
            ],
            "title": "Status",
            "type": "string"
          },
          "target_at": {
            "description": "UTC ISO-8601 ending in Z",
            "format": "date-time",
            "title": "Target At",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "status",
          "started_at",
          "expires_at",
          "target_at",
          "protocol_version",
          "scoring_version",
          "input_mode",
          "schedule",
          "assessment_id",
          "receipt_path"
        ],
        "title": "SessionRecoveryResponse",
        "type": "object"
      },
...
      "StartAssessmentResponse": {
        "properties": {
          "expires_at": {
            "description": "UTC ISO-8601 ending in Z",
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "input_mode": {
            "$ref": "#/components/schemas/InputMode"
          },
          "protocol": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RenderProtocolOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "Task-launch material; null when a replayed session is no longer STARTED"
          },
          "protocol_version": {
            "title": "Protocol Version",
            "type": "string"
          },
          "replayed": {
            "title": "Replayed",
            "type": "boolean"
          },
          "schedule": {
            "$ref": "#/components/schemas/SessionSchedule"
          },
          "scoring_version": {
            "title": "Scoring Version",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "started_at": {
            "description": "UTC ISO-8601 ending in Z",
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "description": "Effective status at server time (expiry is not a GET write)",
            "enum": [
              "STARTED",
              "SUBMITTED",
              "ABANDONED",
              "EXPIRED"
            ],
            "title": "Status",
            "type": "string"
          },
          "target_at": {
            "description": "UTC ISO-8601 ending in Z",
            "format": "date-time",
            "title": "Target At",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "status",
          "started_at",
          "expires_at",
          "target_at",
          "protocol_version",
          "scoring_version",
          "input_mode",
          "schedule",
          "protocol",
          "replayed"
        ],
        "title": "StartAssessmentResponse",
        "type": "object"
      },
      "StartSessionRequest": {
        "additionalProperties": false,
        "description": "All fields required. The client cannot set target time, source, protocol, model or forecast.",
        "properties": {
          "allow_unscheduled": {
            "title": "Allow Unscheduled",
            "type": "boolean"
          },
          "device_changed": {
            "title": "Device Changed",
            "type": "boolean"
          },
          "input_mode": {
            "$ref": "#/components/schemas/InputMode"
          },
          "navigation_assistance": {
            "title": "Navigation Assistance",
            "type": "boolean"
          },
          "start_key": {
            "format": "uuid",
            "title": "Start Key",
            "type": "string"
          }
        },
        "required": [
          "start_key",
          "input_mode",
          "device_changed",
          "navigation_assistance",
          "allow_unscheduled"
        ],
        "title": "StartSessionRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Bearer token from POST /auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Cognitive monitoring prototype API (synthetic/demo data only; not a diagnostic or emergency service). All errors use the shared `ErrorResponse` envelope.",
    "title": "COGNUANCE",
    "version": "backend_contract_v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/patient/assessment-sessions": {
      "post": {
        "description": "Start and freeze one session (protocol, schedule, comparability, forecast) before any task runs.\n\n409 codes: ACTIVE_SESSION_EXISTS, IDEMPOTENCY_CONFLICT, OUTSIDE_SCHEDULE_WINDOW,\nATTEMPT_LIMIT_REACHED, SLOT_COMPLETED.",
        "operationId": "startAssessmentSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartAssessmentResponse"
                }
              }
            },
            "description": "Idempotent replay of the same start_key"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartAssessmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked bearer credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authenticated caller has the wrong role"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Idempotency mismatch, invalid state/transition or stale version"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request body exceeds the limit"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request body is not application/json"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Field or protocol validation failed"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Session",
        "tags": [
          "patient assessments"
        ]
      }
    },
    "/api/v1/patient/assessment-sessions/{session_id}": {
      "get": {
        "description": "Owned session metadata. Task material and raw answers are never re-exposed.",
        "operationId": "getAssessmentSession",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRecoveryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked bearer credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authenticated caller has the wrong role"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Record missing or not accessible to this caller"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Field or protocol validation failed"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Session",
        "tags": [
          "patient assessments"
        ]
      }
    },
    "/api/v1/patient/assessment-sessions/{session_id}/abandon": {
      "post": {
        "description": "STARTED → ABANDONED (+ audit event). Already abandoned/expired → 204; submitted → 409.",
        "operationId": "abandonAssessmentSession",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked bearer credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authenticated caller has the wrong role"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Record missing or not accessible to this caller"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Idempotency mismatch, invalid state/transition or stale version"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Field or protocol validation failed"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Abandon",
        "tags": [
          "patient assessments"
        ]
      }
    },
    "/api/v1/patient/assessment-sessions/{session_id}/submissions": {
      "post": {
        "description": "Phase A (score + durable receipt) then a short bounded phase-B analysis attempt.\n\nA committed receipt is returned even if analysis fails (analysis shows ANALYSIS_ERROR/PENDING).\n409 codes: IDEMPOTENCY_CONFLICT, SESSION_EXPIRED, SESSION_ALREADY_SUBMITTED, SESSION_ABANDONED.",
        "operationId": "submitAssessment",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentReceipt"
                }
              }
            },
            "description": "Idempotent replay (same key and body)"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked bearer credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authenticated caller has the wrong role"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Record missing or not accessible to this caller"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Idempotency mismatch, invalid state/transition or stale version"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request body exceeds the limit"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request body is not application/json"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Field or protocol validation failed"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Infrastructure unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit",
        "tags": [
          "patient assessments"
        ]
      }
    },
    "/api/v1/patient/assessment-status": {
      "get": {
        "description": "Schedule, last receipt and unfinished-session recovery. Read-only (never mutates).",
        "operationId": "getAssessmentStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked bearer credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authenticated caller has the wrong role"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Assessment Status",
        "tags": [
          "patient assessments"
        ]
      }
    },
    "/api/v1/patient/assessments/{assessment_id}/receipt": {
      "get": {
        "description": "Safe persisted receipt with the current analysis availability. GET never processes work.",
        "operationId": "getAssessmentReceipt",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assessment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked bearer credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authenticated caller has the wrong role"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Record missing or not accessible to this caller"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Field or protocol validation failed"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Receipt",
        "tags": [
          "patient assessments"
        ]
      }
    },
...
  }
}